+++ tests/instruments/thorlabs/test_mpc320.py
@@ -2,7 +2,7 @@ import random
 import unittest
 import unittest.mock
 import qmi
-from qmi.core.exceptions import QMI_InstrumentException
+from qmi.core.exceptions import QMI_InstrumentException, QMI_TimeoutException
 from qmi.core.transport import QMI_SerialTransport
 from qmi.instruments.thorlabs import Thorlabs_Mpc320
 from qmi.instruments.thorlabs.apt_protocol import (
@@ -21,9 +21,7 @@ class TestThorlabsMPC320(unittest.TestCase):
             "qmi.instruments.thorlabs.mpc320.create_transport",
             return_value=self._transport_mock,
         ):
-            self._instr: Thorlabs_Mpc320 = qmi.make_instrument(
-                "test_mpc320", Thorlabs_Mpc320, "serial:transport_str"
-            )
+            self._instr: Thorlabs_Mpc320 = qmi.make_instrument("test_mpc320", Thorlabs_Mpc320, "serial:transport_str")
         self._instr.open()
 
     def tearDown(self):
@@ -52,9 +50,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self.assertEqual(idn.serial, expected_idn[2])
         self.assertEqual(idn.version, expected_idn[3])
 
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x05\x00\x00\x00P\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x05\x00\x00\x00P\x01"))
         self._transport_mock.read.assert_has_calls(
             [
                 unittest.mock.call(nbytes=6, timeout=1.0),
@@ -74,9 +70,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         # Assert
         with self.assertRaises(QMI_InstrumentException):
             _ = self._instr.get_idn()
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x05\x00\x00\x00P\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x05\x00\x00\x00P\x01"))
         self._transport_mock.read.assert_has_calls(
             [
                 unittest.mock.call(nbytes=6, timeout=1.0),
@@ -92,9 +86,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.identify()
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x23\x02\x01\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x23\x02\x01\x00\x50\x01"))
 
     def test_enable_channels_enable_four_throws_exception(self):
         """Test enable channel 4, throws exception."""
@@ -113,9 +105,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.enable_channels([1])
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x10\x02\x01\x01\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x10\x02\x01\x01\x50\x01"))
 
     def test_enable_channels_enable_one_and_three_sends_command(self):
         """Test enable channels 1 and 3, sends command to enable channels 1 and 3."""
@@ -125,9 +115,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.enable_channels([1, 3])
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x10\x02\x05\x01\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x10\x02\x05\x01\x50\x01"))
 
     def test_disable_all_channels_sends_command(self):
         """Test disable all channels, sends command to disable all channels."""
@@ -137,9 +125,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.disable_all_channels()
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x10\x02\x00\x01\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x10\x02\x00\x01\x50\x01"))
 
     def test_get_channel_state_for_channel_four_throws_exception(self):
         """Test get state of channel 4, throws exception."""
@@ -162,9 +148,7 @@ class TestThorlabsMPC320(unittest.TestCase):
 
         # Assert
         self.assertEqual(state, AptChannelState.DISABLE)
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x11\x02\x01\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x11\x02\x01\x00\x50\x01"))
 
     def test_get_channel_state_for_enabled_channel_1_sends_command_and_returns_enabled_value(
         self,
@@ -178,9 +162,7 @@ class TestThorlabsMPC320(unittest.TestCase):
 
         # Assert
         self.assertEqual(state, AptChannelState.ENABLE)
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x11\x02\x01\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x11\x02\x01\x00\x50\x01"))
 
     def test_get_channel_state_for_disabled_channel_2_sends_command_and_returns_disabled_value(
         self,
@@ -195,9 +177,7 @@ class TestThorlabsMPC320(unittest.TestCase):
 
         # Assert
         self.assertEqual(state, AptChannelState.DISABLE)
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x11\x02\x01\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x11\x02\x01\x00\x50\x01"))
 
     def test_start_auto_status_update_sends_command(self):
         """Test start automatic status updates, sends command"""
@@ -207,9 +187,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.start_auto_status_update()
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x11\x00\x00\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x11\x00\x00\x00\x50\x01"))
 
     def test_stop_auto_status_update_sends_command(self):
         """Test stop automatic status updates, sends command"""
@@ -219,9 +197,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.stop_auto_status_update()
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x12\x00\x00\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x12\x00\x00\x00\x50\x01"))
 
     def test_home_channel_3_sends_command(self):
         """Test home channel 3, sends command"""
@@ -231,9 +207,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.home_channel(3)
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x43\x04\x04\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x43\x04\x04\x00\x50\x01"))
 
     def test_is_channel_1_homed_when_channel_homed_sends_command_returns_status(self):
         """Test is_channel_homed for channel 1 when channel is homed, send command and return homed value."""
@@ -251,6 +225,31 @@ class TestThorlabsMPC320(unittest.TestCase):
 
         # Assert
         self.assertTrue(state)
+        self._transport_mock.write.assert_not_called()
+        self._transport_mock.read.assert_has_calls(
+            [
+                unittest.mock.call(nbytes=6, timeout=1.0),
+                unittest.mock.call(nbytes=14, timeout=1.0),
+                unittest.mock.call(nbytes=6, timeout=1.0),
+            ]
+        )
+
+    def test_is_channel_1_homed_with_read_timing_out_sends_command_returns_status(self):
+        """Test is_channel_homed for channel 1 when read times out, send command and returns not homed."""
+        # Arrange
+        # first 2 binary strings are responses from the status update command
+        # last string is the homed response
+        self._transport_mock.read.side_effect = [
+            b"\x91\x04\x0e\x00\x00\x81",
+            b"\x00\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
+            QMI_TimeoutException(),
+        ]
+
+        # Act
+        state = self._instr.is_channel_homed(1)
+
+        # Assert
+        self.assertFalse(state)
         self._transport_mock.write.assert_not_called()
         self._transport_mock.read.assert_has_calls(
             [
@@ -309,6 +308,33 @@ class TestThorlabsMPC320(unittest.TestCase):
             ]
         )
 
+    def test_is_channel_1_move_completed_with_read_timing_out_sends_command_returns_status(
+        self,
+    ):
+        """Test is_move_completed for channel 1 when read times out, sends command and returns move not completed."""
+        # Arrange
+        # first 2 binary strings are responses from the status update command
+        # last string is the homed response
+        self._transport_mock.read.side_effect = [
+            b"\x91\x04\x0e\x00\x00\x81",
+            b"\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
+            QMI_TimeoutException,
+        ]
+
+        # Act
+        state = self._instr.is_move_completed(1)
+
+        # Assert
+        self.assertFalse(state)
+        self._transport_mock.write.assert_not_called()
+        self._transport_mock.read.assert_has_calls(
+            [
+                unittest.mock.call(nbytes=6, timeout=1.0),
+                unittest.mock.call(nbytes=14, timeout=1.0),
+                unittest.mock.call(nbytes=6, timeout=1.0),
+            ]
+        )
+
     def test_save_parameter_settings_for_given_command_sends_command(self):
         """Test save_parameter_settings for channel 1 for a specific command, send command."""
         # Arrange
@@ -317,9 +343,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.save_parameter_settings(1, 0x04B6)
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\xb9\x04\x04\x00\xd0\x01\x01\x00\xb6\x04")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\xb9\x04\x04\x00\xd0\x01\x01\x00\xb6\x04"))
 
     def test_get_status_updated_of_channel_1_sends_command_returns_status(
         self,
@@ -339,9 +363,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self.assertEqual(round(status.position), 10)
         self.assertEqual(status.velocity, 0)
         self.assertEqual(status.motor_current, -1)
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x90\x04\x01\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x90\x04\x01\x00\x50\x01"))
         self._transport_mock.read.assert_has_calls(
             [
                 unittest.mock.call(nbytes=6, timeout=1.0),
@@ -357,9 +379,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.jog(1, AptChannelJogDirection.FORWARD)
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x6a\x04\x01\x01\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x6a\x04\x01\x01\x50\x01"))
 
     def test_jog_backward_sends_command(self):
         """Test jog backward for channel 1, sends command."""
@@ -369,9 +389,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self._instr.jog(1, AptChannelJogDirection.BACKWARD)
 
         # Assert
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x6a\x04\x01\x02\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x6a\x04\x01\x02\x50\x01"))
 
     def test_set_polarisation_parameters_sends_command(self):
         """Test set_polarisation_parameters, sends command."""
@@ -393,9 +411,7 @@ class TestThorlabsMPC320(unittest.TestCase):
 
         # Assert
         self._transport_mock.write.assert_called_once_with(
-            bytearray(
-                b"\x30\x05\x0c\x00\xd0\x01\x00\x00\x32\x00\xad\x02\x00\x00\x89\x00\x12\x01"
-            )
+            bytearray(b"\x30\x05\x0c\x00\xd0\x01\x00\x00\x32\x00\xad\x02\x00\x00\x89\x00\x12\x01")
         )
 
     def test_set_polarisation_parameters_with_invalid_velocity_throws_error(self):
@@ -522,9 +538,7 @@ class TestThorlabsMPC320(unittest.TestCase):
         self.assertEqual(params.jog_step1, expected_jog_step_1)
         self.assertEqual(params.jog_step2, expected_jog_step_2)
         self.assertEqual(params.jog_step3, expected_jog_step_3)
-        self._transport_mock.write.assert_called_once_with(
-            bytearray(b"\x31\x05\x00\x00\x50\x01")
-        )
+        self._transport_mock.write.assert_called_once_with(bytearray(b"\x31\x05\x00\x00\x50\x01"))
         self._transport_mock.read.assert_has_calls(
             [
                 unittest.mock.call(nbytes=6, timeout=1.0),